Add services and sections

+++ kernel/src/main/java/fr/unice/polytech/rythmML/kernel/utils/TemporalUtils.java
@@ -8,12 +8,10 @@ public class TemporalUtils {
     }
 
     public static int toTime(final int bar, final int beat, final double division, final int nbBeatPerBar, final int bpm) {
-        //double ticksPerSecond = MIDI_RESOLUTION * 120.0 / bpm;
-        //double ticksPerSecond = MIDI_RESOLUTION  * bpm / 60;
         int pos = bar * nbBeatPerBar * MIDI_RESOLUTION;
         pos += beat * MIDI_RESOLUTION;
         pos += division * MIDI_RESOLUTION;
-        System.out.println("TICK : "+ pos);
-        return (int) (pos);
+        System.out.println("TICK : " + pos);
+        return pos;
     }
 }
